added different ranged enemy types
- created an enemy controller component that handles the enemy behaviour logic and glues the other components together
- created different ranged enemy types
