--- Enemies/Scenes/ranged_enemy_heavy.tscn
@@ -1,155 +1,0 @@
-[gd_scene load_steps=9 format=3 uid="uid://bdx2vl45xi84t"]
-
-[ext_resource type="Script" path="res://Enemies/Scripts/RangedEnemyBase.cs" id="1_mdx8f"]
-[ext_resource type="Resource" uid="uid://de3fgx3nh7wsq" path="res://Resources/RangedEnemyHeavy.tres" id="2_m38g7"]
-[ext_resource type="PackedScene" uid="uid://c4sphn5j4wm3d" path="res://Enemies/Scenes/enemy_bullet.tscn" id="3_0kjri"]
-[ext_resource type="Texture2D" uid="uid://dqx6rpywerr0u" path="res://icon.svg" id="4_4enlh"]
-
-[sub_resource type="RectangleShape2D" id="RectangleShape2D_s2dxq"]
-size = Vector2(14, 34)
-
-[sub_resource type="SpriteFrames" id="SpriteFrames_m7vm3"]
-animations = [{
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"death",
-"speed": 5.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"hurt",
-"speed": 5.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"idle",
-"speed": 5.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"shoot",
-"speed": 5.0
-}]
-
-[sub_resource type="SpriteFrames" id="SpriteFrames_rmubn"]
-animations = [{
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"death",
-"speed": 5.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"hurt",
-"speed": 5.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"idle",
-"speed": 5.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": ExtResource("4_4enlh")
-}],
-"loop": true,
-"name": &"shoot",
-"speed": 5.0
-}]
-
-[sub_resource type="CircleShape2D" id="CircleShape2D_y8g34"]
-radius = 250.0
-
-[node name="RangedEnemyHeavy" type="Area2D" node_paths=PackedStringArray("_spriteBody", "_spriteEye", "_eyeMarker", "_playerDetectionArea", "_wallDetectionRay", "_hurtStaggerTimer", "_shotCooldownTimer", "_debugStateLabel", "_debugHealthLabel") groups=["Enemies"]]
-collision_layer = 32
-collision_mask = 8
-script = ExtResource("1_mdx8f")
-_rangedEnemyStats = ExtResource("2_m38g7")
-_spriteBody = NodePath("spriteBody")
-_spriteEye = NodePath("spriteEye")
-_eyeMarker = NodePath("eyeMarker")
-_playerDetectionArea = NodePath("PlayerDetectionArea")
-_wallDetectionRay = NodePath("WallDetectionRaycast")
-_hurtStaggerTimer = NodePath("HurtStaggerTimer")
-_shotCooldownTimer = NodePath("ShotCooldownTimer")
-_bulletScene = ExtResource("3_0kjri")
-_debugStateLabel = NodePath("StateLabel")
-_debugHealthLabel = NodePath("HealthLabel")
-
-[node name="collider" type="CollisionShape2D" parent="."]
-shape = SubResource("RectangleShape2D_s2dxq")
-debug_color = Color(0.937255, 0, 0.196078, 0.419608)
-
-[node name="spriteBody" type="AnimatedSprite2D" parent="."]
-self_modulate = Color(0.0196078, 0.0196078, 0.0196078, 1)
-scale = Vector2(0.0878907, 0.267578)
-sprite_frames = SubResource("SpriteFrames_m7vm3")
-animation = &"shoot"
-
-[node name="spriteEye" type="AnimatedSprite2D" parent="."]
-modulate = Color(1, 0, 0, 1)
-position = Vector2(0, -11)
-scale = Vector2(0.0488281, 0.048828)
-sprite_frames = SubResource("SpriteFrames_rmubn")
-animation = &"shoot"
-
-[node name="eyeMarker" type="Marker2D" parent="."]
-position = Vector2(0, -11)
-
-[node name="PlayerDetectionArea" type="Area2D" parent="."]
-position = Vector2(0, -11)
-collision_layer = 0
-monitorable = false
-
-[node name="collider" type="CollisionShape2D" parent="PlayerDetectionArea"]
-shape = SubResource("CircleShape2D_y8g34")
-debug_color = Color(0.6, 0.545098, 0, 0.129412)
-
-[node name="WallDetectionRaycast" type="RayCast2D" parent="."]
-self_modulate = Color(0, 0, 0, 1)
-position = Vector2(0, -11)
-target_position = Vector2(-145, 0)
-collision_mask = 6
-
-[node name="HurtStaggerTimer" type="Timer" parent="."]
-
-[node name="ShotCooldownTimer" type="Timer" parent="."]
-
-[node name="RapidFireTimer" type="Timer" parent="."]
-
-[node name="StateLabel" type="Label" parent="."]
-offset_left = -8.0
-offset_top = -30.0
-offset_right = 35.0
-offset_bottom = -17.0
-theme_override_colors/font_color = Color(0, 0, 0, 1)
-theme_override_font_sizes/font_size = 9
-
-[node name="HealthLabel" type="Label" parent="."]
-offset_left = -8.0
-offset_top = -44.0
-offset_right = 35.0
-offset_bottom = -31.0
-theme_override_colors/font_color = Color(0, 0, 0, 1)
-theme_override_font_sizes/font_size = 9
