+++ Enemies/Scenes/ranged_enemy_heavy_shotgunner.tscn
@@ -1,25 +1,25 @@
-[gd_scene load_steps=13 format=3 uid="uid://cswtvyds80uje"]
+[gd_scene load_steps=13 format=3 uid="uid://d3dej0n7nd0so"]
 
-[ext_resource type="Script" path="res://Enemies/Scripts/RangedEnemyHeavyMachineGunner.cs" id="1_itndo"]
-[ext_resource type="Script" path="res://Components/Scripts/HealthComponent.cs" id="1_xwoja"]
-[ext_resource type="Script" path="res://Components/Scripts/HurtboxComponent.cs" id="2_dk7sr"]
-[ext_resource type="Script" path="res://Components/Scripts/EnemyControllerComponent.cs" id="2_inllw"]
-[ext_resource type="Script" path="res://Components/Scripts/PlayerDetectionComponent.cs" id="3_tk62a"]
-[ext_resource type="Script" path="res://Components/Scripts/ShootingComponent.cs" id="4_ukard"]
-[ext_resource type="Script" path="res://Components/Scripts/AnimationComponent.cs" id="5_46qp2"]
-[ext_resource type="Texture2D" uid="uid://dqx6rpywerr0u" path="res://icon.svg" id="6_1sfgy"]
+[ext_resource type="Script" path="res://Components/Scripts/EnemyControllerComponent.cs" id="1_o3boa"]
+[ext_resource type="Script" path="res://Enemies/Scripts/RangedEnemyHeavyShotgunner.cs" id="1_w2m5o"]
+[ext_resource type="Script" path="res://Components/Scripts/HealthComponent.cs" id="2_ua2ig"]
+[ext_resource type="Script" path="res://Components/Scripts/HurtboxComponent.cs" id="3_s107b"]
+[ext_resource type="Script" path="res://Components/Scripts/PlayerDetectionComponent.cs" id="4_gqt0v"]
+[ext_resource type="Script" path="res://Components/Scripts/ShootingComponent.cs" id="5_mrgy3"]
+[ext_resource type="Script" path="res://Components/Scripts/AnimationComponent.cs" id="6_v37to"]
+[ext_resource type="Texture2D" uid="uid://dqx6rpywerr0u" path="res://icon.svg" id="7_swpmd"]
 
-[sub_resource type="RectangleShape2D" id="RectangleShape2D_aud4t"]
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_c6qyb"]
 size = Vector2(14, 34)
 
-[sub_resource type="CircleShape2D" id="CircleShape2D_852hi"]
+[sub_resource type="CircleShape2D" id="CircleShape2D_el1y6"]
 radius = 300.0
 
-[sub_resource type="SpriteFrames" id="SpriteFrames_3cuf2"]
+[sub_resource type="SpriteFrames" id="SpriteFrames_fnitr"]
 animations = [{
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"death",
@@ -27,7 +27,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"hurt",
@@ -35,7 +35,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"idle",
@@ -43,18 +43,18 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"shoot",
 "speed": 5.0
 }]
 
-[sub_resource type="SpriteFrames" id="SpriteFrames_2k87p"]
+[sub_resource type="SpriteFrames" id="SpriteFrames_7p816"]
 animations = [{
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"death",
@@ -62,7 +62,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"hurt",
@@ -70,7 +70,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"idle",
@@ -78,19 +78,18 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("7_swpmd")
 }],
 "loop": true,
 "name": &"shoot",
 "speed": 5.0
 }]
 
-[node name="RangedEnemyHeavyMachineGunner" type="Node2D" node_paths=PackedStringArray("_enemyController")]
-script = ExtResource("1_itndo")
-_enemyController = NodePath("EnemyControllerComponent")
+[node name="RangedEnemyHeavyShotgunner" type="Node2D"]
+script = ExtResource("1_w2m5o")
 
 [node name="EnemyControllerComponent" type="Node" parent="." node_paths=PackedStringArray("_health", "_hurtbox", "_playerDetection", "_shooting", "_animation", "_debugHealthLabel")]
-script = ExtResource("2_inllw")
+script = ExtResource("1_o3boa")
 _enemyType = 1
 _health = NodePath("../HealthComponent")
 _hurtbox = NodePath("../HurtboxComponent")
@@ -100,18 +99,18 @@ _animation = NodePath("../AnimationComponent")
 _debugHealthLabel = NodePath("../HealthLabel")
 
 [node name="HealthComponent" type="Node" parent="."]
-script = ExtResource("1_xwoja")
+script = ExtResource("2_ua2ig")
 CurrentHealth = 60
 MaxHealth = 60
 
 [node name="HurtboxComponent" type="Area2D" parent="." node_paths=PackedStringArray("_hurtStaggerTimer") groups=["Enemies"]]
 collision_layer = 32
 collision_mask = 8
-script = ExtResource("2_dk7sr")
+script = ExtResource("3_s107b")
 _hurtStaggerTimer = NodePath("HurtStaggerTimer")
 
 [node name="collider" type="CollisionShape2D" parent="HurtboxComponent"]
-shape = SubResource("RectangleShape2D_aud4t")
+shape = SubResource("RectangleShape2D_c6qyb")
 debug_color = Color(0.937255, 0, 0.196078, 0.419608)
 
 [node name="HurtStaggerTimer" type="Timer" parent="HurtboxComponent"]
@@ -119,7 +118,7 @@ wait_time = 0.4
 one_shot = true
 
 [node name="PlayerDetectionComponent" type="Node2D" parent="." node_paths=PackedStringArray("_playerDetectionArea", "_playerDetectionRay")]
-script = ExtResource("3_tk62a")
+script = ExtResource("4_gqt0v")
 _playerDetectionArea = NodePath("PlayerDetectionArea")
 _playerDetectionRay = NodePath("PlayerDetectionRay")
 
@@ -129,7 +128,7 @@ collision_layer = 0
 monitorable = false
 
 [node name="collider" type="CollisionShape2D" parent="PlayerDetectionComponent/PlayerDetectionArea"]
-shape = SubResource("CircleShape2D_852hi")
+shape = SubResource("CircleShape2D_el1y6")
 debug_color = Color(0.806739, 0.415537, 0.0885346, 0.129412)
 
 [node name="PlayerDetectionRay" type="RayCast2D" parent="PlayerDetectionComponent"]
@@ -139,16 +138,16 @@ target_position = Vector2(230, 0)
 collision_mask = 6
 
 [node name="ShootingComponent" type="Node2D" parent="." node_paths=PackedStringArray("_muzzle", "_shotCooldownTimer", "_reloadTimer")]
-script = ExtResource("4_ukard")
-_weaponType = 2
-_shootingCooldownTime = 0.05
-_reloadTime = 3.5
-_magazineSize = 60
-_bulletDamage = 2
+script = ExtResource("5_mrgy3")
+_weaponType = 1
+_shootingCooldownTime = 1.3
+_reloadTime = 2.1
+_magazineSize = 4
+_bulletDamage = 5
 _bulletKnockback = 35.0
-_bulletsPerShot = 1
-_bulletSwayAngle = 15.0
-_bulletSpeed = 450.0
+_bulletsPerShot = 5
+_bulletSwayAngle = 5.0
+_bulletSpeed = 380.0
 _muzzle = NodePath("muzzle")
 _shotCooldownTimer = NodePath("ShotCooldownTimer")
 _reloadTimer = NodePath("ReloadTimer")
@@ -161,20 +160,20 @@ position = Vector2(0, -11)
 [node name="ReloadTimer" type="Timer" parent="ShootingComponent"]
 
 [node name="AnimationComponent" type="Node" parent="." node_paths=PackedStringArray("Sprite")]
-script = ExtResource("5_46qp2")
+script = ExtResource("6_v37to")
 Sprite = NodePath("../spriteBody")
 
 [node name="spriteBody" type="AnimatedSprite2D" parent="."]
 self_modulate = Color(0.0196078, 0.0196078, 0.0196078, 1)
 scale = Vector2(0.0878907, 0.267578)
-sprite_frames = SubResource("SpriteFrames_3cuf2")
+sprite_frames = SubResource("SpriteFrames_fnitr")
 animation = &"shoot"
 
 [node name="spriteEye" type="AnimatedSprite2D" parent="."]
 modulate = Color(1, 0, 0, 1)
 position = Vector2(0, -11)
 scale = Vector2(0.0488281, 0.048828)
-sprite_frames = SubResource("SpriteFrames_2k87p")
+sprite_frames = SubResource("SpriteFrames_7p816")
 animation = &"shoot"
 
 [node name="StateLabel" type="Label" parent="."]
