+++ Enemies/Scenes/ranged_enemy_light_pistol.tscn
@@ -1,25 +1,25 @@
-[gd_scene load_steps=13 format=3 uid="uid://cswtvyds80uje"]
+[gd_scene load_steps=13 format=3 uid="uid://dnkd8nwcia0in"]
 
-[ext_resource type="Script" path="res://Enemies/Scripts/RangedEnemyHeavyMachineGunner.cs" id="1_itndo"]
-[ext_resource type="Script" path="res://Components/Scripts/HealthComponent.cs" id="1_xwoja"]
-[ext_resource type="Script" path="res://Components/Scripts/HurtboxComponent.cs" id="2_dk7sr"]
-[ext_resource type="Script" path="res://Components/Scripts/EnemyControllerComponent.cs" id="2_inllw"]
-[ext_resource type="Script" path="res://Components/Scripts/PlayerDetectionComponent.cs" id="3_tk62a"]
-[ext_resource type="Script" path="res://Components/Scripts/ShootingComponent.cs" id="4_ukard"]
-[ext_resource type="Script" path="res://Components/Scripts/AnimationComponent.cs" id="5_46qp2"]
-[ext_resource type="Texture2D" uid="uid://dqx6rpywerr0u" path="res://icon.svg" id="6_1sfgy"]
+[ext_resource type="Script" path="res://Enemies/Scripts/RangedEnemyLightPistol.cs" id="1_djtq4"]
+[ext_resource type="Script" path="res://Components/Scripts/EnemyControllerComponent.cs" id="2_jd8vp"]
+[ext_resource type="Script" path="res://Components/Scripts/HealthComponent.cs" id="3_donhs"]
+[ext_resource type="Script" path="res://Components/Scripts/HurtboxComponent.cs" id="4_u8cat"]
+[ext_resource type="Script" path="res://Components/Scripts/PlayerDetectionComponent.cs" id="5_ost8k"]
+[ext_resource type="Script" path="res://Components/Scripts/ShootingComponent.cs" id="6_lo5xk"]
+[ext_resource type="Script" path="res://Components/Scripts/AnimationComponent.cs" id="7_765es"]
+[ext_resource type="Texture2D" uid="uid://dqx6rpywerr0u" path="res://icon.svg" id="8_wqfqs"]
 
-[sub_resource type="RectangleShape2D" id="RectangleShape2D_aud4t"]
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_oa4wd"]
 size = Vector2(14, 34)
 
-[sub_resource type="CircleShape2D" id="CircleShape2D_852hi"]
-radius = 300.0
+[sub_resource type="CircleShape2D" id="CircleShape2D_xy5a3"]
+radius = 250.0
 
-[sub_resource type="SpriteFrames" id="SpriteFrames_3cuf2"]
+[sub_resource type="SpriteFrames" id="SpriteFrames_nf60s"]
 animations = [{
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"death",
@@ -27,7 +27,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"hurt",
@@ -35,7 +35,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"idle",
@@ -43,18 +43,18 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"shoot",
 "speed": 5.0
 }]
 
-[sub_resource type="SpriteFrames" id="SpriteFrames_2k87p"]
+[sub_resource type="SpriteFrames" id="SpriteFrames_1cgpy"]
 animations = [{
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"death",
@@ -62,7 +62,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"hurt",
@@ -70,7 +70,7 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"idle",
@@ -78,19 +78,19 @@ animations = [{
 }, {
 "frames": [{
 "duration": 1.0,
-"texture": ExtResource("6_1sfgy")
+"texture": ExtResource("8_wqfqs")
 }],
 "loop": true,
 "name": &"shoot",
 "speed": 5.0
 }]
 
-[node name="RangedEnemyHeavyMachineGunner" type="Node2D" node_paths=PackedStringArray("_enemyController")]
-script = ExtResource("1_itndo")
+[node name="RangedEnemyLightPistol" type="Node2D" node_paths=PackedStringArray("_enemyController")]
+script = ExtResource("1_djtq4")
 _enemyController = NodePath("EnemyControllerComponent")
 
 [node name="EnemyControllerComponent" type="Node" parent="." node_paths=PackedStringArray("_health", "_hurtbox", "_playerDetection", "_shooting", "_animation", "_debugHealthLabel")]
-script = ExtResource("2_inllw")
+script = ExtResource("2_jd8vp")
 _enemyType = 1
 _health = NodePath("../HealthComponent")
 _hurtbox = NodePath("../HurtboxComponent")
@@ -100,26 +100,26 @@ _animation = NodePath("../AnimationComponent")
 _debugHealthLabel = NodePath("../HealthLabel")
 
 [node name="HealthComponent" type="Node" parent="."]
-script = ExtResource("1_xwoja")
-CurrentHealth = 60
-MaxHealth = 60
+script = ExtResource("3_donhs")
+CurrentHealth = 30
+MaxHealth = 30
 
 [node name="HurtboxComponent" type="Area2D" parent="." node_paths=PackedStringArray("_hurtStaggerTimer") groups=["Enemies"]]
 collision_layer = 32
 collision_mask = 8
-script = ExtResource("2_dk7sr")
+script = ExtResource("4_u8cat")
 _hurtStaggerTimer = NodePath("HurtStaggerTimer")
 
 [node name="collider" type="CollisionShape2D" parent="HurtboxComponent"]
-shape = SubResource("RectangleShape2D_aud4t")
+shape = SubResource("RectangleShape2D_oa4wd")
 debug_color = Color(0.937255, 0, 0.196078, 0.419608)
 
 [node name="HurtStaggerTimer" type="Timer" parent="HurtboxComponent"]
-wait_time = 0.4
+wait_time = 0.9
 one_shot = true
 
 [node name="PlayerDetectionComponent" type="Node2D" parent="." node_paths=PackedStringArray("_playerDetectionArea", "_playerDetectionRay")]
-script = ExtResource("3_tk62a")
+script = ExtResource("5_ost8k")
 _playerDetectionArea = NodePath("PlayerDetectionArea")
 _playerDetectionRay = NodePath("PlayerDetectionRay")
 
@@ -129,8 +129,8 @@ collision_layer = 0
 monitorable = false
 
 [node name="collider" type="CollisionShape2D" parent="PlayerDetectionComponent/PlayerDetectionArea"]
-shape = SubResource("CircleShape2D_852hi")
-debug_color = Color(0.806739, 0.415537, 0.0885346, 0.129412)
+shape = SubResource("CircleShape2D_xy5a3")
+debug_color = Color(0.6, 0.545098, 0, 0.129412)
 
 [node name="PlayerDetectionRay" type="RayCast2D" parent="PlayerDetectionComponent"]
 self_modulate = Color(0, 0, 0, 1)
@@ -139,16 +139,15 @@ target_position = Vector2(230, 0)
 collision_mask = 6
 
 [node name="ShootingComponent" type="Node2D" parent="." node_paths=PackedStringArray("_muzzle", "_shotCooldownTimer", "_reloadTimer")]
-script = ExtResource("4_ukard")
-_weaponType = 2
-_shootingCooldownTime = 0.05
-_reloadTime = 3.5
-_magazineSize = 60
-_bulletDamage = 2
-_bulletKnockback = 35.0
+script = ExtResource("6_lo5xk")
+_shootingCooldownTime = 0.5
+_reloadTime = 1.5
+_magazineSize = 6
+_bulletDamage = 5
+_bulletKnockback = 25.0
 _bulletsPerShot = 1
-_bulletSwayAngle = 15.0
-_bulletSpeed = 450.0
+_bulletSwayAngle = 2.3
+_bulletSpeed = 350.0
 _muzzle = NodePath("muzzle")
 _shotCooldownTimer = NodePath("ShotCooldownTimer")
 _reloadTimer = NodePath("ReloadTimer")
@@ -161,20 +160,20 @@ position = Vector2(0, -11)
 [node name="ReloadTimer" type="Timer" parent="ShootingComponent"]
 
 [node name="AnimationComponent" type="Node" parent="." node_paths=PackedStringArray("Sprite")]
-script = ExtResource("5_46qp2")
-Sprite = NodePath("../spriteBody")
+script = ExtResource("7_765es")
+Sprite = NodePath("")
 
 [node name="spriteBody" type="AnimatedSprite2D" parent="."]
 self_modulate = Color(0.0196078, 0.0196078, 0.0196078, 1)
 scale = Vector2(0.0878907, 0.267578)
-sprite_frames = SubResource("SpriteFrames_3cuf2")
+sprite_frames = SubResource("SpriteFrames_nf60s")
 animation = &"shoot"
 
 [node name="spriteEye" type="AnimatedSprite2D" parent="."]
 modulate = Color(1, 0, 0, 1)
 position = Vector2(0, -11)
 scale = Vector2(0.0488281, 0.048828)
-sprite_frames = SubResource("SpriteFrames_2k87p")
+sprite_frames = SubResource("SpriteFrames_1cgpy")
 animation = &"shoot"
 
 [node name="StateLabel" type="Label" parent="."]
