--- Enemies/Scripts/RangedEnemyBase.cs
@@ -1,273 +1,0 @@
-using Godot;
-using ElGatoProject.Resources;
-using ElGatoProject.Singletons;
-
-namespace ElGatoProject.Enemies.Scripts;
-
-[GlobalClass]
-public partial class RangedEnemyBase : Area2D
-{
-	[Export] private RangedEnemyStats _rangedEnemyStats;
-	[Export] private AnimatedSprite2D _spriteBody, _spriteEye;
-	[Export] private Marker2D _eyeMarker;
-	[Export] private Area2D _playerDetectionArea;
-	[Export] private RayCast2D _wallDetectionRay;
-	[Export] private Timer _hurtStaggerTimer;
-	[Export] private Timer _shotCooldownTimer;
-	[Export] private Timer _rapidFireTimer;
-	[Export] private PackedScene _bulletScene;
-	[Export] private Label _debugStateLabel;
-	[Export] private Label _debugHealthLabel;
-	
-	private bool _hurtStatus, _playerInRange, _onCooldown, _rapidFireCooldown;
-	private Area2D _playerProjectile;
-	private Node2D _player;
-	private int _bulletCount;
-	
-	public override void _Ready()
-	{
-		SubscribeToEvents();
-		
-		if (_hurtStaggerTimer != null)
-		{
-			_hurtStaggerTimer.OneShot = true;
-			_hurtStaggerTimer.SetWaitTime(_rangedEnemyStats.HurtStaggerTime);
-			_hurtStaggerTimer.Timeout += HurtStaggerTimerTimedOut;
-		}
-
-		if (_shotCooldownTimer != null)
-		{
-			_shotCooldownTimer.OneShot = true;
-            _shotCooldownTimer.SetWaitTime(_rangedEnemyStats.AttackCooldownTime);
-            _shotCooldownTimer.Timeout += ShotCooldownTimerTimedOut;
-		}
-		
-		if (_rapidFireTimer != null && _rangedEnemyStats.RangedEnemyType == Utility.RangedEnemyType.RangedEnemyMachineGun)
-		{
-			_rapidFireTimer.OneShot = true;
-			_rapidFireTimer.SetWaitTime(_rangedEnemyStats.RapidFireTime);
-			_rapidFireTimer.Timeout += RapidFireTimerTimedOut;
-		}
-		
-		_playerDetectionArea.AreaEntered += PlayerEnteredDetectionRange;
-		_playerDetectionArea.AreaExited += PlayerExitedDetectionRange;
-
-		_wallDetectionRay.Enabled = false;
-		
-		// For debug only, remove later
-		_debugStateLabel.SetText("idle");
-		_debugHealthLabel.SetText("HP: "+ _rangedEnemyStats.Health);
-	}
-
-	private void SubscribeToEvents()
-	{
-		EventsBus.Instance.OnAttackHit += HitByAttack;
-	}
-
-	private void UnsubscribeFromEvents()
-	{
-		EventsBus.Instance.OnAttackHit -= HitByAttack;
-	}
-	
-	// Detecting when player enters or exits detection range
-	private void PlayerEnteredDetectionRange(Area2D area)
-	{
-		if (area.IsInGroup("Players"))
-		{
-			_player = area;
-			_playerInRange = true;
-			_wallDetectionRay.Enabled = true;
-			_wallDetectionRay.TargetPosition = ToLocal(_player.GlobalPosition);
-			
-			ResetBulletCount();
-		}
-	}
-
-	private void PlayerExitedDetectionRange(Area2D area)
-	{
-		if (area.IsInGroup("Players"))
-		{
-			_playerInRange = false;
-			_wallDetectionRay.Enabled = false;
-			
-			_debugStateLabel.SetText("Idle");
-		}
-	}
-	
-	// Enemy behaviour
-	private void EnemyBehaviour()
-	{
-		// Use raycast to detect if line of sight to player is blocked by wall
-		if (_playerInRange)
-		{
-			_wallDetectionRay.TargetPosition = ToLocal(_player.GlobalPosition);
-
-			if (!_wallDetectionRay.IsColliding() && !_hurtStatus)
-			{
-				Shoot();
-			}
-			else if (_hurtStatus)
-			{
-				_rangedEnemyStats.State = Utility.EntityState.Hurt;
-				ResetBulletCount();
-			}
-			else
-			{
-				_rangedEnemyStats.State = Utility.EntityState.Idle;
-			}
-		}
-		else
-		{
-			_rangedEnemyStats.State = _hurtStatus ? Utility.EntityState.Hurt : Utility.EntityState.Idle;
-		}
-	}
-
-	private void Shoot()
-	{
-		switch (_rangedEnemyStats.RangedEnemyType)
-		{
-			case Utility.RangedEnemyType.RangedEnemyLight:
-			case Utility.RangedEnemyType.RangedEnemyHeavy:
-				if (!_onCooldown)
-				{
-					_debugStateLabel.SetText("Shooting");
-					SpawnBullets(false);
-					_onCooldown = true;
-					_shotCooldownTimer.Start();
-				}
-				break;
-			
-			case Utility.RangedEnemyType.RangedEnemyMachineGun:
-				if (!_rapidFireCooldown && !_onCooldown)
-				{
-					_debugStateLabel.SetText("Shooting");
-					SpawnBullets(true);
-					_rapidFireCooldown = true;
-					_rapidFireTimer.Start();
-					_bulletCount++;
-
-					if (_bulletCount >= _rangedEnemyStats.BulletsPerShot)
-					{
-						_onCooldown = true;
-						_shotCooldownTimer.Start();
-						_bulletCount = 0;
-					}
-				}
-				break;
-		}
-	}
-	
-	private void SpawnBullets(bool rapidFire)
-	{
-		if (!rapidFire)
-		{
-			for (int i = 0; i < _rangedEnemyStats.BulletsPerShot; i++)
-			{
-				InstantiateBullet();
-			}
-		}
-		else
-		{
-			InstantiateBullet();
-		}
-	}
-	
-	private void InstantiateBullet()
-	{
-		var bulletInstance = (EnemyBullet)_bulletScene.Instantiate();
-		
-		// Set properties for the bullet
-		bulletInstance.Target = GlobalPosition.DirectionTo(_player.GlobalPosition);
-		bulletInstance.RotationDegrees = Globals.Instance.Rng.RandfRange(-_rangedEnemyStats.BulletAngle, _rangedEnemyStats.BulletAngle);
-		bulletInstance.BulletSpeed = _rangedEnemyStats.BulletSpeed;
-		bulletInstance.Knockback = _rangedEnemyStats.Knockback;
-		bulletInstance.BulletDespawnTimeSeconds = _rangedEnemyStats.BulletDespawnTimeSeconds;
-		bulletInstance.BulletDamage = _rangedEnemyStats.BulletDamage;
-		bulletInstance.GlobalPosition = _eyeMarker.GlobalPosition;
-		
-		// Add bullet instance to scene
-		GetTree().Root.AddChild(bulletInstance);
-	}
-
-	private void RapidFireTimerTimedOut()
-	{
-		_rapidFireCooldown = false;
-	}
-
-	private void ResetBulletCount()
-	{
-		if (_rangedEnemyStats.RangedEnemyType == Utility.RangedEnemyType.RangedEnemyMachineGun)
-		{
-			_bulletCount = 0;
-		}
-	}
-	
-	private void ShotCooldownTimerTimedOut()
-	{
-		_onCooldown = false;
-	}
-	
-	// Getting hit by player bullets
-	private void HitByAttack(Area2D attackArea, Area2D entityArea, int attackDamage, float knockback, Vector2 attackVelocity)
-	{
-		if (!attackArea.IsInGroup("PlayerProjectiles"))
-			return;
-		if (entityArea != this)
-			return;
-		
-		_hurtStatus = true;
-		_hurtStaggerTimer.Start();
-		_rangedEnemyStats.TakeDamage(attackDamage);
-		
-		_debugHealthLabel.SetText("HP: "+ _rangedEnemyStats.Health);
-	}
-	
-	private void HurtStaggerTimerTimedOut()
-	{
-		_hurtStatus = false;
-	}
-	
-	private void Animations()
-	{
-		switch (_rangedEnemyStats.State)
-		{
-			case Utility.EntityState.Idle:
-				_spriteBody.Play("idle");
-				_spriteEye.Play("idle");
-				
-				_debugStateLabel.SetText("idle");
-				break;
-			case Utility.EntityState.Shoot:
-				_spriteBody.Play("shoot");
-				_spriteEye.Play("shoot");
-				
-				_debugStateLabel.SetText("shoot");
-				break;
-			case Utility.EntityState.Hurt:
-				_spriteBody.Play("hurt");
-				_spriteEye.Play("hurt");
-				
-				_debugStateLabel.SetText("hurt");
-				break;
-			case Utility.EntityState.Death:
-				_spriteBody.Play("death");
-				_spriteEye.Play("death");
-				
-				_debugStateLabel.SetText("death");
-				QueueFree();
-				break;
-		}
-	}
-
-	public override void _Process(double delta)
-	{
-		EnemyBehaviour();
-		Animations();
-	}
-	
-	// Unsubscribe to events when exiting scene tree to prevent memory leaks
-	public override void _ExitTree()
-	{
-		UnsubscribeFromEvents();
-	}
-}
